Tidy up recipe files
