+++ front-admin/src/modules/recipes/RecipeCreate.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Create, Datagrid, SimpleForm, TextInput, LongTextInput } from 'admin-on-rest';
+import AutoComplete from 'material-ui/AutoComplete';
+
+export const RecipeCreate = (props) => (
+    <Create {...props}>
+        <SimpleForm>
+            <TextInput source="title" />
+            <LongTextInput source="body" />
+        </SimpleForm>
+    </Create>
+);
