+++ front-admin/src/modules/recipes/RecipeEdit.js
@@ -1,15 +1,6 @@
 import React from 'react';
-import { List, Edit, Create, Datagrid, TextField, EditButton, DisabledInput, LongTextInput, ReferenceInput, required, SimpleForm, TextInput, AutocompleteInput } from 'admin-on-rest';
+import { Edit, DisabledInput, ReferenceInput, required, SimpleForm, TextInput, AutocompleteInput } from 'admin-on-rest';
 import AutoComplete from 'material-ui/AutoComplete';
-
-export const RecipeList = (props) => (
-    <List {...props}>
-        <Datagrid>
-            <TextField source="name" />
-            <EditButton />
-        </Datagrid>
-    </List>
-);
 
 const RecipeTitle = ({ record }) => {
     return <span>Recipe {record ? `"${record.name}"` : ''}</span>;
@@ -30,11 +21,3 @@ export const RecipeEdit = (props) => (
     </Edit>
 );
 
-export const RecipeCreate = (props) => (
-    <Create {...props}>
-        <SimpleForm>
-            <TextInput source="title" />
-            <LongTextInput source="body" />
-        </SimpleForm>
-    </Create>
-);
