+++ front-admin/src/modules/recipes/RecipeList.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { List, Datagrid, TextField, EditButton} from 'admin-on-rest';
+
+export const RecipeList = (props) => (
+    <List {...props}>
+        <Datagrid>
+            <TextField source="name" />
+            <EditButton />
+        </Datagrid>
+    </List>
+);
