+++ front-admin/src/modules/recipes/index.js
@@ -1,0 +1,3 @@
+export * from './RecipeEdit'
+export * from './RecipeList'
+export * from './RecipeCreate'
